+++ app/assets/scripts/minireel/services.js
@@ -1033,9 +1033,9 @@ function( angular , c6uilib , cryptojs , c6Defines  ) {
 
             _private.fetchGettyThumbs = function(imageid) {
                 return {
-                    small: 'http://embed-cdn.gettyimages.com/xt/' + imageid +
+                    small: '//embed-cdn.gettyimages.com/xt/' + imageid +
                         '.jpg?v=1&g=fs1|0|DV|33|651&s=1',
-                    large: 'http://embed-cdn.gettyimages.com/xt/' + imageid +
+                    large: '//embed-cdn.gettyimages.com/xt/' + imageid +
                         '.jpg?v=1&g=fs1|0|DV|33|651&s=1'
                 };
             };
@@ -1316,7 +1316,7 @@ function( angular , c6uilib , cryptojs , c6Defines  ) {
             // This function uses GettyImages' oEmbed API endpoint to fetch the embed code for an
             // image
             _private.getGettyEmbedInfo = function(imageid) {
-                var request = 'http://embed.gettyimages.com/oembed?' +
+                var request = '//embed.gettyimages.com/oembed?' +
                     'url=' + encodeURIComponent('http://gty.im/' + imageid);
                 return $http.get(request, {cache: true}).
                     then(function(json) {
@@ -2227,6 +2227,10 @@ function( angular , c6uilib , cryptojs , c6Defines  ) {
                         return 'AOL On';
                     case 'yahoo':
                         return 'Yahoo! Screen';
+                    case 'flickr':
+                        return 'Flickr';
+                    case 'getty':
+                        return 'gettyimages';
                     case 'adUnit':
                         return undefined;
                     default:
@@ -2378,6 +2382,9 @@ function( angular , c6uilib , cryptojs , c6Defines  ) {
                         href: imageHrefValue(),
                         width: embedValue('width'),
                         height: embedValue('height'),
+                        source: function(data) {
+                            return camelSource(data.service);
+                        },
                         thumbs: imageThumbsValue()
                     },
                     youtube: {